@@ -9,8 +9,9 @@
 		<AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
 		<AppendRuntimeIdentifierToOutputPath>false</AppendRuntimeIdentifierToOutputPath>
 		<GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-		<LangVersion>11</LangVersion>
+		<LangVersion>14</LangVersion>
 		<Nullable>disable</Nullable>
+		<RestoreSources>$(RestoreSources);C:\GitHub\.nupkgs</RestoreSources>
 	</PropertyGroup>
 	<PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
 		<DebugType>portable</DebugType>
